[모듈 이름 수정 및 JSONEncoder 수정] - 모듈 이름을 명확히 하기 위해 view.py -> seller_view.py 로 변경 - JSONEncoder 속성을 개별 view마다 지정해주는 것이 아닌 app 자체 속성으로 변경
